Conditional on tag deletion

+++ openshift/pipeline/src/groovy/Clean.groovy
@@ -51,10 +51,13 @@ config.app.namespaces.each {String key, Map env->
 
         // Delete ImageStreamTags associated to a single PR (most others)
         Map ret = oc(['get','istag', """-o=jsonpath='{.items[?(@.tag.name=="build-pr-${config.app.git.changeId}")].metadata.name}'""", '--ignore-not-found=true', '-n', "${env.namespace}"])
-        tags = ret['out'].split()
-        println "Tags: ${tags}"
-        tags.each { tag ->
-            oc(['delete', 'istag', "${tag}", '--ignore-not-found=true', '-n', "${env.namespace}"])
+        out = ret['out']
+        if (out != ''){
+            tags = out.split()
+            println "Tags: ${tags}"
+            tags.each { tag ->
+                oc(['delete', 'istag', "${tag}", '--ignore-not-found=true', '-n', "${env.namespace}"])
+            }
         }
     }
 }
